debug: introduce electronDebugActions.ts
fixes #8806
