+++ src/vs/workbench/parts/debug/electron-browser/debugEditorContribution.ts
@@ -14,8 +14,8 @@ import { KeyCode } from 'vs/base/common/keyCodes';
 import keyboard = require('vs/base/browser/keyboardEvent');
 import editorbrowser = require('vs/editor/browser/editorBrowser');
 import editorcommon = require('vs/editor/common/editorCommon');
-import { DebugHoverWidget } from 'vs/workbench/parts/debug/browser/debugHover';
-import debugactions = require('vs/workbench/parts/debug/electron-browser/debugActions');
+import { DebugHoverWidget } from 'vs/workbench/parts/debug/electron-browser/debugHover';
+import debugactions = require('vs/workbench/parts/debug/browser/debugActions');
 import debug = require('vs/workbench/parts/debug/common/debug');
 import { IWorkspaceContextService } from 'vs/workbench/services/workspace/common/contextService';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
