+++ src/vs/workbench/parts/debug/electron-browser/debugViews.ts
@@ -18,8 +18,8 @@ import splitview = require('vs/base/browser/ui/splitview/splitview');
 import viewlet = require('vs/workbench/browser/viewlet');
 import debug = require('vs/workbench/parts/debug/common/debug');
 import { StackFrame, Expression, Variable, ExceptionBreakpoint, FunctionBreakpoint } from 'vs/workbench/parts/debug/common/debugModel';
-import viewer = require('vs/workbench/parts/debug/browser/debugViewer');
-import debugactions = require('vs/workbench/parts/debug/electron-browser/debugActions');
+import viewer = require('vs/workbench/parts/debug/electron-browser/debugViewer');
+import debugactions = require('vs/workbench/parts/debug/browser/debugActions');
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
